Update lens to use double.

+++ KsApi/models/lenses/ProjectStatsEnvelope.CumulativeLenses.swift
@@ -22,7 +22,7 @@ extension ProjectStatsEnvelope.Cumulative {
         percentRaised: $1.percentRaised, pledged: $1.pledged) }
     )
 
-    public static let percentRaised = Lens<ProjectStatsEnvelope.Cumulative, Float>(
+    public static let percentRaised = Lens<ProjectStatsEnvelope.Cumulative, Double>(
       view: { $0.percentRaised },
       set: { .init(averagePledge: $1.averagePledge, backersCount: $1.backersCount, goal: $1.goal,
         percentRaised: $0, pledged: $1.pledged) }
